Refactor RPC gateway handling to use SSR-specific mappings and improve remote config retrieval
