+++ apps/earn-protocol/app/server-handlers/sumr-decay-factor/index.ts
@@ -6,7 +6,7 @@ import { createPublicClient, http } from 'viem'
 import { base } from 'viem/chains'
 
 import { backendSDK } from '@/app/server-handlers/sdk/sdk-backend-client'
-import { SDKChainIdToRpcGatewayMap } from '@/constants/networks-list'
+import { SDKChainIdToSSRRpcGatewayMap } from '@/helpers/rpc-gateway-ssr'
 
 export interface SumrDecayFactorData {
   address: string
@@ -29,7 +29,7 @@ export const getSumrDecayFactor = async (addresses: string[]): Promise<SumrDecay
 
     const publicClient = createPublicClient({
       chain: base,
-      transport: http(SDKChainIdToRpcGatewayMap[SDKChainId.BASE]),
+      transport: http(await SDKChainIdToSSRRpcGatewayMap[SDKChainId.BASE]),
     })
 
     let sumrToken
